# airbyte-api/src/main/openapi/api.yaml
openapi: 3.0.0
info:
  title: Airbyte Public API
  version: 1.0.0
servers:
  - url: "http://localhost:8006/v1"
    description: Airbyte API v1
paths:
...
  /connections:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionCreateRequest"
            examples:
              Connection Creation Request Example:
                value:
                  sourceId: 95e66a59-8045-4307-9678-63bc3c9b8c93
                  destinationId: e478de0d-a3a0-475c-b019-25f7dd29e281
                  name: Postgres-to-Bigquery
        required: true
      tags:
        - Connections
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionResponse"
              examples:
                Connection Creation Response Example:
                  value:
                    connectionId: 9924bcd0-99be-453d-ba47-c2c9766f7da5
          description: Successful operation
        "400":
          description: Invalid data
        "403":
          description: Not allowed
      operationId: createConnection
      summary: Create a connection
    get:
      tags:
        - Connections
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionsResponse"
          description: Successful operation
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: listConnections
      summary: List connections
      parameters:
        - name: workspaceIds
          description: The UUIDs of the workspaces you wish to list connections for. Empty list will retrieve all allowed workspaces.
          schema:
            type: array
            items:
              format: uuid
              type: string
          in: query
          required: false
        - name: includeDeleted
          description: Include deleted connections in the returned results.
          schema:
            default: false
            type: boolean
          in: query
          required: false
        - name: limit
          description: Set the limit on the number of Connections returned. The default is 20.
          schema:
            format: int32
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          in: query
        - name: offset
          description: Set the offset to start at when returning Connections. The default is 0
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          in: query
  /connections/{connectionId}:
    get:
      tags:
        - Connections
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionResponse"
              examples:
                Connection Get Response Example:
                  value:
                    workspaceId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: Postgres To Snowflake
                    sourceId: 9924bcd0-99be-453d-ba47-c2c9766f7da5
                    destinationId: 744cc0ed-7f05-4949-9e60-2a814f90c035
          description: Get a Connection by the id in the path.
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: getConnection
      summary: Get Connection details
    patch:
      tags:
        - Connections
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionPatchRequest"
            examples:
              Connection Update Request Example:
                value:
                  sourceId: 95e66a59-8045-4307-9678-63bc3c9b8c93
                  destinationId: e478de0d-a3a0-475c-b019-25f7dd29e281
                  name: Postgres-to-Bigquery
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionResponse"
              examples:
                Connection Get Response Example:
                  value:
                    workspaceId: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
                    name: Postgres To Snowflake
                    sourceId: 9924bcd0-99be-453d-ba47-c2c9766f7da5
                    destinationId: 744cc0ed-7f05-4949-9e60-2a814f90c035
          description: Update a Connection by the id in the path.
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: patchConnection
      summary: Update Connection details
    delete:
      tags:
        - Connections
      responses:
        "204":
          description: The resource was deleted successfully
        "403":
          description: Not allowed
        "404":
          description: Not found
      operationId: deleteConnection
      summary: Delete a Connection
    parameters:
      - name: connectionId
        schema:
          format: UUID
          type: string
        in: path
        required: true
...
components:
...
  schemas:
...
    ConnectionCreateRequest:
      required:
        - sourceId
        - destinationId
      type: object
      properties:
        name:
          description: Optional name of the connection
          type: string
        sourceId:
          format: uuid
          type: string
        destinationId:
          format: uuid
          type: string
        configurations:
          $ref: "#/components/schemas/StreamConfigurations"
        schedule:
          $ref: "#/components/schemas/ConnectionSchedule"
        dataResidency:
          $ref: "#/components/schemas/GeographyEnum"
        namespaceDefinition:
          $ref: "#/components/schemas/NamespaceDefinitionEnum"
        namespaceFormat:
          type: string
          description: Used when namespaceDefinition is 'custom_format'. If blank then behaves like namespaceDefinition = 'destination'. If "${SOURCE_NAMESPACE}" then behaves like namespaceDefinition = 'source'.
          default: null
          example: "${SOURCE_NAMESPACE}"
        prefix:
          type: string
          description: Prefix that will be prepended to the name of each stream when it is written to the destination (ex. “airbyte_” causes “projects” => “airbyte_projects”).
        nonBreakingSchemaUpdatesBehavior:
          $ref: "#/components/schemas/NonBreakingSchemaUpdatesBehaviorEnum"
        status:
          $ref: "#/components/schemas/ConnectionStatusEnum"
    ConnectionPatchRequest:
      type: object
      properties:
        name:
          description: Optional name of the connection
          type: string
        configurations:
          $ref: "#/components/schemas/StreamConfigurations"
        schedule:
          $ref: "#/components/schemas/ConnectionSchedule"
        dataResidency:
          $ref: "#/components/schemas/GeographyEnumNoDefault"
        namespaceDefinition:
          $ref: "#/components/schemas/NamespaceDefinitionEnumNoDefault"
        namespaceFormat:
          type: string
          description: Used when namespaceDefinition is 'custom_format'. If blank then behaves like namespaceDefinition = 'destination'. If "${SOURCE_NAMESPACE}" then behaves like namespaceDefinition = 'source'.
          default: null
          example: "${SOURCE_NAMESPACE}"
        prefix:
          type: string
          description: Prefix that will be prepended to the name of each stream when it is written to the destination (ex. “airbyte_” causes “projects” => “airbyte_projects”).
        nonBreakingSchemaUpdatesBehavior:
          $ref: "#/components/schemas/NonBreakingSchemaUpdatesBehaviorEnumNoDefault"
        status:
          $ref: "#/components/schemas/ConnectionStatusEnum"
...
    ConnectionResponse:
      title: Root Type for ConnectionResponse
      description: Provides details of a single connection.
      type: object
      required:
        - connectionId
        - name
        - sourceId
        - destinationId
        - workspaceId
        - status
        - schedule
        - dataResidency
        - configurations
      properties:
        connectionId:
          format: UUID
          type: string
        name:
          type: string
        sourceId:
          format: UUID
          type: string
        destinationId:
          format: UUID
          type: string
        workspaceId:
          format: UUID
          type: string
        status:
          $ref: "#/components/schemas/ConnectionStatusEnum"
        schedule:
          $ref: "#/components/schemas/ConnectionScheduleResponse"
        dataResidency:
          $ref: "#/components/schemas/GeographyEnum"
        nonBreakingSchemaUpdatesBehavior:
          $ref: "#/components/schemas/NonBreakingSchemaUpdatesBehaviorEnum"
        namespaceDefinition:
          $ref: "#/components/schemas/NamespaceDefinitionEnum"
        namespaceFormat:
          type: string
        prefix:
          type: string
        configurations:
          $ref: "#/components/schemas/StreamConfigurations"
    ConnectionSchedule:
      description: schedule for when the the connection should run, per the schedule type
      type: object
      required:
        - scheduleType
      properties:
        scheduleType:
          $ref: "#/components/schemas/ScheduleTypeEnum"
        cronExpression:
          type: string
    ScheduleTypeEnum:
      type: string
      enum:
        - manual
        - cron
    ConnectionScheduleResponse:
      description: schedule for when the the connection should run, per the schedule type
      type: object
      required:
        - scheduleType
      properties:
        scheduleType:
          $ref: "#/components/schemas/ScheduleTypeWithBasicEnum"
        cronExpression:
          type: string
        basicTiming:
          type: string
    ScheduleTypeWithBasicEnum:
      type: string
      enum:
        - manual
        - cron
        - basic
    GeographyEnum:
      type: string
      enum:
        - auto
        - us
        - eu
      default: "auto"
    GeographyEnumNoDefault:
      type: string
      enum:
        - auto
        - us
        - eu
    ConnectionStatusEnum:
      type: string
      enum:
        - active
        - inactive
        - deprecated
    NamespaceDefinitionEnum:
      type: string
      description: Define the location where the data will be stored in the destination
      enum:
        - source
        - destination
        - custom_format
      default: "destination"
    NonBreakingSchemaUpdatesBehaviorEnum:
      type: string
      description: Set how Airbyte handles syncs when it detects a non-breaking schema change in the source
      enum:
        - ignore # do nothing if we detect a schema change
        - disable_connection # disable the connection, pausing the sync
        - propagate_columns # Auto propagate schema change for existing streams
        - propagate_fully # Auto propagate schema change for all streams including the new ones
      default: "ignore"
    NamespaceDefinitionEnumNoDefault:
      type: string
      description: Define the location where the data will be stored in the destination
      enum:
        - source
        - destination
        - custom_format
    NonBreakingSchemaUpdatesBehaviorEnumNoDefault:
      type: string
      description: Set how Airbyte handles syncs when it detects a non-breaking schema change in the source
      enum:
        - ignore # do nothing if we detect a schema change
        - disable_connection # disable the connection, pausing the sync
        - propagate_columns # Auto propagate schema change for existing streams
        - propagate_fully # Auto propagate schema change for all streams including the new ones
...
    ConnectionsResponse:
      title: Root Type for ConnectionsResponse
      description: ""
      required:
        - data
      type: object
      properties:
        previous:
          type: string
        next:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/ConnectionResponse"
          default: []
      example:
        next: "https://api.airbyte.com/v1/connections?limit=5&offset=10"
        previous: "https://api.airbyte.com/v1/connections?limit=5&offset=0"
        data:
          - name: "test-connection"
          - connection_id: 18dccc91-0ab1-4f72-9ed7-0b8fc27c5826
          - sourceId: 49237019-645d-47d4-b45b-5eddf97775ce
          - destinationId: al312fs-0ab1-4f72-9ed7-0b8fc27c5826
          - schedule:
              scheduleType: manual
          - status: active
          - dataResidency: auto
...
    StreamConfiguration:
      description: Configurations for a single stream.
      type: object
      required:
        - name
      properties:
        name:
          type: string
        syncMode:
          $ref: "#/components/schemas/ConnectionSyncModeEnum"
        cursorField:
          description: Path to the field that will be used to determine if a record is new or modified since the last sync. This field is REQUIRED if `sync_mode` is `incremental` unless there is a default.
          type: array
          items:
            type: string
        primaryKey:
          description: Paths to the fields that will be used as primary key. This field is REQUIRED if `destination_sync_mode` is `*_dedup` unless it is already supplied by the source schema.
          type: array
          items:
            type: array
            items:
              type: string
    StreamConfigurations:
      description: A list of configured stream options for a connection.
      type: object
      properties:
        streams:
          type: array
          items:
            $ref: "#/components/schemas/StreamConfiguration"
...
    ConnectionSyncModeEnum:
      enum:
        - full_refresh_overwrite
        - full_refresh_append
        - incremental_append
        - incremental_deduped_history
...